+++ folioflex/dashboard/app.py
@@ -210,4 +210,4 @@ def update_auth_status(pathname):
 
 if __name__ == "__main__":
     custom_logger.set_log_level("DEBUG", module_prefix="pages")
-    app.run_server(debug=True, host="0.0.0.0")
+    app.run(debug=True, host="0.0.0.0")